requirements.txt
# Web Framework
Flask==2.2.3

# WSGI HTTP Server (for production)
gunicorn==20.1.0

# Docker integration
docker==6.1.3

# AI Generation
google-generativeai==0.3.1

# Utilities
python-dotenv==1.0.0
psutil==5.9.5
pyyaml==6.0.1
requests==2.31.0
Werkzeug==2.2.3

# Security
cryptography==41.0.5